pyproject.toml
[tool.poetry]
name = "steer"
version = "0.1.0"
description = ""
authors = ["José Coelho <16445494+jcoelho93@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
click = "8.1.7"
pydantic = "^2.7.1"
questionary = "^2.0.1"
jsonpath-ng = "^1.6.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
coverage = "^7.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
